Update README.md
This script was forked from one created by nytrobound in order to get around the 'Client with the currently selected authenticator does not support any combination of challenges that will satisfy the CA' error. This error emerged in January, 2018, due to Let's Encrypt disabling TLS-SNI-01 challenges because of a security issue, but then releasing Certbot 0.21.0 (that allows HTTP-01 challenges).